Windows: Scheduled tasks - code review fixes
