+++ volatility3/framework/plugins/windows/scheduled_tasks.py
@@ -11,8 +11,6 @@ import io
 import itertools
 import logging
 import struct
-import traceback
-import unittest
 from typing import Dict, Iterator, List, Optional, Tuple, Union
 
 from volatility3.framework import exceptions, interfaces, renderers
@@ -50,6 +48,16 @@ class TimeMode(enum.Enum):
     DaysInWeeksInMonths = "Days In Weeks in Months"
 
     Unknown = "Unknown"
+
+
+TIME_MODE_DESCRIPTION = {
+    TimeMode.Once: "Once",
+    TimeMode.Daily: "Daily",
+    TimeMode.Weekly: "Weekly",
+    TimeMode.DaysInMonths: "Days In Months",
+    TimeMode.DaysInWeeksInMonths: "Days In Weeks In Months",
+    TimeMode.Unknown: "Unknown",
+}
 
 
 class ActionType(enum.Enum):
@@ -193,13 +201,8 @@ class _ScheduledTasksReader(io.BytesIO):
         filetime = self.decode_filetime()
         if filetime is None:
             return None
+
         return is_localized, filetime
-
-    def read_filetime(self) -> Optional[datetime.datetime]:
-        return datetime.datetime.now()
-
-    def seek_relative(self, offset) -> int:
-        return self.seek(offset, io.SEEK_CUR)
 
     def read_bool(self, aligned=False) -> Optional[bool]:
         try:
@@ -561,13 +564,14 @@ class _JobSchedule:
         repetition_duration_secs = reader.read_u4()
         execution_time_limit_secs = reader.read_u4()
         mode_index = reader.read_u4()
-        try:
-            if mode_index is not None:
-                mode = list(TimeMode)[mode_index]
-            else:
-                mode = None
-        except IndexError:
-            mode = TimeMode.Unknown
+        if mode_index is not None:
+            try:
+                mode = TimeMode(mode_index)
+            except ValueError:
+                mode = TimeMode.Unknown
+        else:
+            mode = None
+
         data1 = reader.read_u2()
         data2 = reader.read_u2()
         data3 = reader.read_u2()
@@ -1412,385 +1416,3 @@ class ScheduledTasks(interfaces.plugins.PluginInterface, timeliner.TimeLinerInte
                 for indent, entry in self._generator()
             ),
         )
-
-
-class TestActionsDecoding(unittest.TestCase):
-    def test_decode_exe_action(self):
-        # fmt: off
-        buf = struct.pack(
-            "512B",
-            *[
-                0x03, 0x00, 0x16, 0x00, 0x00, 0x00, 0x4c, 0x00,
-                0x6f, 0x00, 0x63, 0x00, 0x61, 0x00, 0x6c, 0x00,
-                0x53, 0x00, 0x79, 0x00, 0x73, 0x00, 0x74, 0x00,
-                0x65, 0x00, 0x6d, 0x00, 0x66, 0x66, 0x00, 0x00,
-                0x00, 0x00, 0x6e, 0x00, 0x00, 0x00, 0x25, 0x00,
-                0x77, 0x00, 0x69, 0x00, 0x6e, 0x00, 0x64, 0x00,
-                0x69, 0x00, 0x72, 0x00, 0x25, 0x00, 0x5c, 0x00,
-                0x73, 0x00, 0x79, 0x00, 0x73, 0x00, 0x74, 0x00,
-                0x65, 0x00, 0x6d, 0x00, 0x33, 0x00, 0x32, 0x00,
-                0x5c, 0x00, 0x57, 0x00, 0x69, 0x00, 0x6e, 0x00,
-                0x64, 0x00, 0x6f, 0x00, 0x77, 0x00, 0x73, 0x00,
-                0x50, 0x00, 0x6f, 0x00, 0x77, 0x00, 0x65, 0x00,
-                0x72, 0x00, 0x53, 0x00, 0x68, 0x00, 0x65, 0x00,
-                0x6c, 0x00, 0x6c, 0x00, 0x5c, 0x00, 0x76, 0x00,
-                0x31, 0x00, 0x2e, 0x00, 0x30, 0x00, 0x5c, 0x00,
-                0x70, 0x00, 0x6f, 0x00, 0x77, 0x00, 0x65, 0x00,
-                0x72, 0x00, 0x73, 0x00, 0x68, 0x00, 0x65, 0x00,
-                0x6c, 0x00, 0x6c, 0x00, 0x2e, 0x00, 0x65, 0x00,
-                0x78, 0x00, 0x65, 0x00, 0x62, 0x01, 0x00, 0x00,
-                0x2d, 0x00, 0x45, 0x00, 0x78, 0x00, 0x65, 0x00,
-                0x63, 0x00, 0x75, 0x00, 0x74, 0x00, 0x69, 0x00,
-                0x6f, 0x00, 0x6e, 0x00, 0x50, 0x00, 0x6f, 0x00,
-                0x6c, 0x00, 0x69, 0x00, 0x63, 0x00, 0x79, 0x00,
-                0x20, 0x00, 0x55, 0x00, 0x6e, 0x00, 0x72, 0x00,
-                0x65, 0x00, 0x73, 0x00, 0x74, 0x00, 0x72, 0x00,
-                0x69, 0x00, 0x63, 0x00, 0x74, 0x00, 0x65, 0x00,
-                0x64, 0x00, 0x20, 0x00, 0x2d, 0x00, 0x4e, 0x00,
-                0x6f, 0x00, 0x6e, 0x00, 0x49, 0x00, 0x6e, 0x00,
-                0x74, 0x00, 0x65, 0x00, 0x72, 0x00, 0x61, 0x00,
-                0x63, 0x00, 0x74, 0x00, 0x69, 0x00, 0x76, 0x00,
-                0x65, 0x00, 0x20, 0x00, 0x2d, 0x00, 0x4e, 0x00,
-                0x6f, 0x00, 0x50, 0x00, 0x72, 0x00, 0x6f, 0x00,
-                0x66, 0x00, 0x69, 0x00, 0x6c, 0x00, 0x65, 0x00,
-                0x20, 0x00, 0x2d, 0x00, 0x57, 0x00, 0x69, 0x00,
-                0x6e, 0x00, 0x64, 0x00, 0x6f, 0x00, 0x77, 0x00,
-                0x53, 0x00, 0x74, 0x00, 0x79, 0x00, 0x6c, 0x00,
-                0x65, 0x00, 0x20, 0x00, 0x48, 0x00, 0x69, 0x00,
-                0x64, 0x00, 0x64, 0x00, 0x65, 0x00, 0x6e, 0x00,
-                0x20, 0x00, 0x22, 0x00, 0x26, 0x00, 0x20, 0x00,
-                0x25, 0x00, 0x77, 0x00, 0x69, 0x00, 0x6e, 0x00,
-                0x64, 0x00, 0x69, 0x00, 0x72, 0x00, 0x25, 0x00,
-                0x5c, 0x00, 0x73, 0x00, 0x79, 0x00, 0x73, 0x00,
-                0x74, 0x00, 0x65, 0x00, 0x6d, 0x00, 0x33, 0x00,
-                0x32, 0x00, 0x5c, 0x00, 0x57, 0x00, 0x69, 0x00,
-                0x6e, 0x00, 0x64, 0x00, 0x6f, 0x00, 0x77, 0x00,
-                0x73, 0x00, 0x50, 0x00, 0x6f, 0x00, 0x77, 0x00,
-                0x65, 0x00, 0x72, 0x00, 0x53, 0x00, 0x68, 0x00,
-                0x65, 0x00, 0x6c, 0x00, 0x6c, 0x00, 0x5c, 0x00,
-                0x76, 0x00, 0x31, 0x00, 0x2e, 0x00, 0x30, 0x00,
-                0x5c, 0x00, 0x4d, 0x00, 0x6f, 0x00, 0x64, 0x00,
-                0x75, 0x00, 0x6c, 0x00, 0x65, 0x00, 0x73, 0x00,
-                0x5c, 0x00, 0x53, 0x00, 0x6d, 0x00, 0x62, 0x00,
-                0x53, 0x00, 0x68, 0x00, 0x61, 0x00, 0x72, 0x00,
-                0x65, 0x00, 0x5c, 0x00, 0x44, 0x00, 0x69, 0x00,
-                0x73, 0x00, 0x61, 0x00, 0x62, 0x00, 0x6c, 0x00,
-                0x65, 0x00, 0x55, 0x00, 0x6e, 0x00, 0x75, 0x00,
-                0x73, 0x00, 0x65, 0x00, 0x64, 0x00, 0x53, 0x00,
-                0x6d, 0x00, 0x62, 0x00, 0x31, 0x00, 0x2e, 0x00,
-                0x70, 0x00, 0x73, 0x00, 0x31, 0x00, 0x20, 0x00,
-                0x2d, 0x00, 0x53, 0x00, 0x63, 0x00, 0x65, 0x00,
-                0x6e, 0x00, 0x61, 0x00, 0x72, 0x00, 0x69, 0x00,
-                0x6f, 0x00, 0x20, 0x00, 0x43, 0x00, 0x6c, 0x00,
-                0x69, 0x00, 0x65, 0x00, 0x6e, 0x00, 0x74, 0x00,
-                0x22, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ]
-        )
-
-        try:
-            actions = ActionSet.decode(buf)  # type: ignore
-            self.assertEqual(len(actions), 1)
-            self.assertEqual(actions[0].action_type, ActionType.Exe)
-        except Exception:
-            self.fail(
-                "ActionDecoder.decode should not raise exception:\n%s"
-                % traceback.format_exc()
-            )
-
-
-class TestTriggersDecoding(unittest.TestCase):
-    def test_decode_all_triggers(self):
-        """
-        Tests decoding a set of all triggers that can be constructed via the
-        Task Scheduler GUI interface. Ensures that the correct number of bytes
-        is being consumed for each trigger structure.
-        """
-        buf = struct.pack(
-            "1808B",
-            # fmt: off
-            *[
-                0x17, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0xda, 0xaf, 0x8d, 0x09, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0xda, 0xaf, 0x8d, 0x09, 0x00, 0x00, 0x00,
-                0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-                0x38, 0x21, 0x41, 0x42, 0x48, 0x48, 0x48, 0x48,
-                0xa0, 0x12, 0xa0, 0xa4, 0x48, 0x48, 0x48, 0x48,
-                0x0e, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x41, 0x00, 0x75, 0x00, 0x74, 0x00, 0x68, 0x00,
-                0x6f, 0x00, 0x72, 0x00, 0x00, 0x00, 0x48, 0x48,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x00, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0x00, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0x01, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x1c, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x01, 0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05,
-                0x15, 0x00, 0x00, 0x00, 0x69, 0xce, 0x28, 0x2a,
-                0xce, 0xd8, 0x1f, 0x77, 0x37, 0x9c, 0xe2, 0x44,
-                0xf4, 0x01, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x40, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x44, 0x00, 0x45, 0x00, 0x53, 0x00, 0x4b, 0x00,
-                0x54, 0x00, 0x4f, 0x00, 0x50, 0x00, 0x2d, 0x00,
-                0x45, 0x00, 0x33, 0x00, 0x38, 0x00, 0x38, 0x00,
-                0x44, 0x00, 0x38, 0x00, 0x50, 0x00, 0x5c, 0x00,
-                0x41, 0x00, 0x64, 0x00, 0x6d, 0x00, 0x69, 0x00,
-                0x6e, 0x00, 0x69, 0x00, 0x73, 0x00, 0x74, 0x00,
-                0x72, 0x00, 0x61, 0x00, 0x74, 0x00, 0x6f, 0x00,
-                0x72, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x2c, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
-                0x80, 0xf4, 0x03, 0x00, 0xff, 0xff, 0xff, 0xff,
-                0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0xdd, 0xdd, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x01, 0x07, 0x0a, 0x00, 0x00, 0x00, 0x09, 0x00,
-                0x80, 0x48, 0x11, 0xf8, 0x36, 0x1a, 0xdb, 0x01,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x2e, 0xe2, 0x01, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0xc2, 0x31, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0xaa, 0xaa, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0xda, 0xaf, 0x8d, 0x09, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0xda, 0xaf, 0x8d, 0x09, 0x00, 0x00, 0x00,
-                0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x01, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0xff, 0xff, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0xda, 0xaf, 0x8d, 0x09, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0xda, 0xaf, 0x8d, 0x09, 0x00, 0x00, 0x00,
-                0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0xee, 0xee, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0xda, 0xaf, 0x8d, 0x09, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0xda, 0xaf, 0x8d, 0x09, 0x00, 0x00, 0x00,
-                0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0xcc, 0xcc, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x65, 0x00, 0x78, 0x00, 0x65, 0x00,
-                0x22, 0x00, 0x20, 0x00, 0x53, 0x00, 0x74, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x84, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x3c, 0x00, 0x51, 0x00, 0x75, 0x00, 0x65, 0x00,
-                0x72, 0x00, 0x79, 0x00, 0x4c, 0x00, 0x69, 0x00,
-                0x73, 0x00, 0x74, 0x00, 0x3e, 0x00, 0x3c, 0x00,
-                0x51, 0x00, 0x75, 0x00, 0x65, 0x00, 0x72, 0x00,
-                0x79, 0x00, 0x20, 0x00, 0x49, 0x00, 0x64, 0x00,
-                0x3d, 0x00, 0x22, 0x00, 0x30, 0x00, 0x22, 0x00,
-                0x20, 0x00, 0x50, 0x00, 0x61, 0x00, 0x74, 0x00,
-                0x68, 0x00, 0x3d, 0x00, 0x22, 0x00, 0x49, 0x00,
-                0x6e, 0x00, 0x74, 0x00, 0x65, 0x00, 0x72, 0x00,
-                0x6e, 0x00, 0x65, 0x00, 0x74, 0x00, 0x20, 0x00,
-                0x45, 0x00, 0x78, 0x00, 0x70, 0x00, 0x6c, 0x00,
-                0x6f, 0x00, 0x72, 0x00, 0x65, 0x00, 0x72, 0x00,
-                0x22, 0x00, 0x3e, 0x00, 0x3c, 0x00, 0x53, 0x00,
-                0x65, 0x00, 0x6c, 0x00, 0x65, 0x00, 0x63, 0x00,
-                0x74, 0x00, 0x20, 0x00, 0x50, 0x00, 0x61, 0x00,
-                0x74, 0x00, 0x68, 0x00, 0x3d, 0x00, 0x22, 0x00,
-                0x49, 0x00, 0x6e, 0x00, 0x74, 0x00, 0x65, 0x00,
-                0x72, 0x00, 0x6e, 0x00, 0x65, 0x00, 0x74, 0x00,
-                0x20, 0x00, 0x45, 0x00, 0x78, 0x00, 0x70, 0x00,
-                0x6c, 0x00, 0x6f, 0x00, 0x72, 0x00, 0x65, 0x00,
-                0x72, 0x00, 0x22, 0x00, 0x3e, 0x00, 0x2a, 0x00,
-                0x5b, 0x00, 0x53, 0x00, 0x79, 0x00, 0x73, 0x00,
-                0x74, 0x00, 0x65, 0x00, 0x6d, 0x00, 0x5b, 0x00,
-                0x45, 0x00, 0x76, 0x00, 0x65, 0x00, 0x6e, 0x00,
-                0x74, 0x00, 0x49, 0x00, 0x44, 0x00, 0x3d, 0x00,
-                0x32, 0x00, 0x5d, 0x00, 0x5d, 0x00, 0x3c, 0x00,
-                0x2f, 0x00, 0x53, 0x00, 0x65, 0x00, 0x6c, 0x00,
-                0x65, 0x00, 0x63, 0x00, 0x74, 0x00, 0x3e, 0x00,
-                0x3c, 0x00, 0x2f, 0x00, 0x51, 0x00, 0x75, 0x00,
-                0x65, 0x00, 0x72, 0x00, 0x79, 0x00, 0x3e, 0x00,
-                0x3c, 0x00, 0x2f, 0x00, 0x51, 0x00, 0x75, 0x00,
-                0x65, 0x00, 0x72, 0x00, 0x79, 0x00, 0x4c, 0x00,
-                0x69, 0x00, 0x73, 0x00, 0x74, 0x00, 0x3e, 0x00,
-                0x00, 0x00, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x88, 0x88, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x77, 0x77, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00,
-                0x01, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x01, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0x77, 0x77, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0x00, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0x01, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x1c, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x01, 0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05,
-                0x15, 0x00, 0x00, 0x00, 0x69, 0xce, 0x28, 0x2a,
-                0xce, 0xd8, 0x1f, 0x77, 0x37, 0x9c, 0xe2, 0x44,
-                0xf4, 0x01, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x40, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x44, 0x00, 0x45, 0x00, 0x53, 0x00, 0x4b, 0x00,
-                0x54, 0x00, 0x4f, 0x00, 0x50, 0x00, 0x2d, 0x00,
-                0x45, 0x00, 0x33, 0x00, 0x38, 0x00, 0x38, 0x00,
-                0x44, 0x00, 0x38, 0x00, 0x50, 0x00, 0x5c, 0x00,
-                0x41, 0x00, 0x64, 0x00, 0x6d, 0x00, 0x69, 0x00,
-                0x6e, 0x00, 0x69, 0x00, 0x73, 0x00, 0x74, 0x00,
-                0x72, 0x00, 0x61, 0x00, 0x74, 0x00, 0x6f, 0x00,
-                0x72, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x77, 0x77, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00,
-                0x01, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x01, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0x77, 0x77, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00,
-                0x01, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0x00, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0x01, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x1c, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x01, 0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05,
-                0x15, 0x00, 0x00, 0x00, 0x69, 0xce, 0x28, 0x2a,
-                0xce, 0xd8, 0x1f, 0x77, 0x37, 0x9c, 0xe2, 0x44,
-                0xf4, 0x01, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x40, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x44, 0x00, 0x45, 0x00, 0x53, 0x00, 0x4b, 0x00,
-                0x54, 0x00, 0x4f, 0x00, 0x50, 0x00, 0x2d, 0x00,
-                0x45, 0x00, 0x33, 0x00, 0x38, 0x00, 0x38, 0x00,
-                0x44, 0x00, 0x38, 0x00, 0x50, 0x00, 0x5c, 0x00,
-                0x41, 0x00, 0x64, 0x00, 0x6d, 0x00, 0x69, 0x00,
-                0x6e, 0x00, 0x69, 0x00, 0x73, 0x00, 0x74, 0x00,
-                0x72, 0x00, 0x61, 0x00, 0x74, 0x00, 0x6f, 0x00,
-                0x72, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-            ]
-            # fmt: on
-        )
-        triggers = TriggerSet.decode(buf)
-        self.assertIsNotNone(triggers)
-
-    def test_decode_triggers(self):
-        # fmt: off
-        buf = struct.pack(
-            "320B",
-            *[
-                0x17, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0xB9, 0x61, 0x1A, 0xA8, 0xB9, 0x61, 0x1A,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0xB9, 0x61, 0x1A, 0xA8, 0xB9, 0x61, 0x1A,
-                0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-                0x08, 0xA1, 0x40, 0x42, 0x48, 0x48, 0x48, 0x48,
-                0x7A, 0x7F, 0x59, 0xDC, 0x48, 0x48, 0x48, 0x48,
-                0x22, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x49, 0x00, 0x6E, 0x00, 0x74, 0x00, 0x65, 0x00,
-                0x72, 0x00, 0x61, 0x00, 0x63, 0x00, 0x74, 0x00,
-                0x69, 0x00, 0x76, 0x00, 0x65, 0x00, 0x55, 0x00,
-                0x73, 0x00, 0x65, 0x00, 0x72, 0x00, 0x73, 0x00,
-                0x00, 0x00, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x00, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0x00, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-                0x05, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x0C, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05,
-                0x04, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x2C, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x00, 0x00, 0x00, 0x00, 0xFF, 0xFF, 0xFF, 0xFF,
-                0x80, 0x51, 0x01, 0x00, 0xFF, 0xFF, 0xFF, 0xFF,
-                0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0xAA, 0xAA, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0xB9, 0x61, 0x1A, 0xA8, 0xB9, 0x61, 0x1A,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0xB9, 0x61, 0x1A, 0xA8, 0xB9, 0x61, 0x1A,
-                0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-                0x2C, 0x01, 0x00, 0x00, 0xFF, 0xFF, 0xFF, 0xFF,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0xC1, 0xD9, 0x04,
-                0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x48, 0x48, 0x48, 0x48,
-                0x01, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48, 0x48,
-            ]
-        )
-        # fmt: on
-        triggers = TriggerSet.decode(buf)
-        self.assertIsNotNone(triggers)
-        if not triggers:
-            return
-        self.assertGreater(len(triggers.triggers), 0)
